Cleaned up program.cs/refactor
